images/uinput-device-plugin/python/requirements.txt
python-uinput
msgpack